# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Goodbase Public API",
    "version": "1.2.0",
    "description": "Goodbase public API for GoodOS apps. All public endpoints require an API key using X-Goodbase-API-Key or Authorization: Bearer."
  },
  "servers": [
    {
      "url": "https://base.goodos.app/api/v1",
      "description": "GoodOS Production Public API"
    }
  ],
  "security": [
    {
      "GoodOSApiKey": []
    },
    {
      "GoodOSBearer": []
    }
  ],
  "components": {
    "securitySchemes": {
      "GoodOSApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Goodbase-API-Key"
      },
      "GoodOSBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "GoodBaseSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "goodos_session"
      }
    },
...
  },
  "paths": {
...
    "/api/goodspeech/v1/speech": {
      "post": {
        "summary": "Generate speech through GoodBase",
        "description": "Authenticated GoodSpeech generation through the private Kokoro service. Provider credentials remain server-side and voice cloning is unavailable.",
        "servers": [
          {
            "url": "https://base.goodos.app"
          }
        ],
        "security": [
          {
            "GoodBaseSession": []
          },
          {
            "GoodOSBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["text", "voice"],
                "properties": {
                  "text": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 2000
                  },
                  "voice": {
                    "type": "object"
                  },
                  "style": {
                    "type": "string"
                  },
                  "tone": {
                    "type": "string"
                  },
                  "intensity": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 100
                  },
                  "contextualExpressiveness": {
                    "type": "boolean"
                  },
                  "language": {
                    "type": "string",
                    "enum": ["en-us", "en-gb", "es", "fr-fr", "hi", "it", "pt-br"],
                    "default": "en-us"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Base64-encoded audio generated successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "429": {
            "description": "Per-user generation limit exceeded"
          },
          "502": {
            "description": "Speech provider unavailable"
          },
          "503": {
            "description": "Speech provider is not configured"
          }
        }
      }
    },
...
    "/api/goodspeech/v1/voices/{voiceId}/speech": {
      "post": {
        "summary": "Generate watermarked speech with a private voice",
        "servers": [{ "url": "https://base.goodos.app" }],
        "security": [{ "GoodBaseSession": [] }, { "GoodOSBearer": [] }],
        "parameters": [{ "in": "path", "name": "voiceId", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["text"], "properties": { "text": { "type": "string", "maxLength": 2000 } } } } } },
        "responses": { "200": { "description": "Watermarked WAV", "content": { "audio/wav": { "schema": { "type": "string", "format": "binary" } } } }, "404": { "description": "Voice not found" }, "410": { "description": "Voice revoked" }, "429": { "description": "Rate or quota exceeded" } }
      },
      "delete": {
        "summary": "Revoke a private voice and delete its reference sample",
        "servers": [{ "url": "https://base.goodos.app" }],
        "security": [{ "GoodBaseSession": [] }, { "GoodOSBearer": [] }],
        "parameters": [{ "in": "path", "name": "voiceId", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Voice revoked and private sample deleted" }, "404": { "description": "Voice not found" } }
      }
    },
...
    "/api/goodspeech/v1/agents/bootstrap": {
      "get": {
        "summary": "Load owned voice agents and analytics",
        "description": "Returns owner-scoped agent configurations, grounded knowledge counts, automated test counts, conversation analytics, and realtime capabilities.",
        "servers": [{ "url": "https://base.goodos.app" }],
        "security": [{ "GoodBaseSession": [] }, { "GoodOSBearer": [] }],
        "responses": { "200": { "description": "Voice agent workspace" }, "401": { "description": "Authentication required" }, "403": { "description": "GoodSpeech entitlement required" } }
      }
    },
    "/api/goodspeech/v1/agents": {
      "post": {
        "summary": "Create a governed GoodSpeech voice agent",
        "description": "Creates an owner-scoped voice agent with a bounded system prompt, approved Kokoro voice, language, retention period, and allowlisted tools.",
        "servers": [{ "url": "https://base.goodos.app" }],
        "security": [{ "GoodBaseSession": [] }, { "GoodOSBearer": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["name", "systemPrompt", "greeting"], "properties": { "name": { "type": "string", "maxLength": 80 }, "systemPrompt": { "type": "string", "maxLength": 8000 }, "greeting": { "type": "string", "maxLength": 500 }, "voice": { "type": "string" }, "language": { "type": "string" }, "tools": { "type": "array", "items": { "type": "string", "enum": ["knowledge_search", "current_time", "usage_summary", "handoff"] } }, "retentionDays": { "type": "integer", "minimum": 0, "maximum": 365 } } } } } },
        "responses": { "201": { "description": "Agent created" }, "400": { "description": "Invalid agent configuration" }, "401": { "description": "Authentication required" } }
      }
    },
    "/api/goodspeech/v1/agents/{agentId}/knowledge": {
      "get": {
        "summary": "List an owned agent's knowledge sources",
        "servers": [{ "url": "https://base.goodos.app" }],
        "security": [{ "GoodBaseSession": [] }, { "GoodOSBearer": [] }],
        "parameters": [{ "in": "path", "name": "agentId", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Knowledge inventory" }, "404": { "description": "Agent not found" } }
      },
      "post": {
        "summary": "Index a grounded knowledge source",
        "servers": [{ "url": "https://base.goodos.app" }],
        "security": [{ "GoodBaseSession": [] }, { "GoodOSBearer": [] }],
        "parameters": [{ "in": "path", "name": "agentId", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "201": { "description": "Knowledge source indexed" }, "400": { "description": "Invalid or oversized source" }, "404": { "description": "Agent not found" } }
      }
    },
    "/api/goodspeech/v1/agents/{agentId}/sessions": {
      "post": {
        "summary": "Start a persistent voice agent session",
        "servers": [{ "url": "https://base.goodos.app" }],
        "security": [{ "GoodBaseSession": [] }, { "GoodOSBearer": [] }],
        "parameters": [{ "in": "path", "name": "agentId", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "201": { "description": "Session started with the configured greeting and voice" }, "404": { "description": "Active agent not found" }, "429": { "description": "Rate limit exceeded" } }
      }
    },
    "/api/goodspeech/v1/agents/sessions/{sessionId}/turns": {
      "post": {
        "summary": "Run a grounded conversational turn",
        "description": "Persists the user and assistant messages, executes at most one allowlisted tool, emits citations, and records response latency.",
        "servers": [{ "url": "https://base.goodos.app" }],
        "security": [{ "GoodBaseSession": [] }, { "GoodOSBearer": [] }],
        "parameters": [{ "in": "path", "name": "sessionId", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "201": { "description": "Grounded response and voice settings" }, "400": { "description": "Invalid message" }, "404": { "description": "Session not found" }, "409": { "description": "Session already ended" } }
      }
    },
    "/api/goodspeech/v1/agents/sessions/{sessionId}/interrupt": {
      "post": {
        "summary": "Record realtime barge-in",
        "servers": [{ "url": "https://base.goodos.app" }],
        "security": [{ "GoodBaseSession": [] }, { "GoodOSBearer": [] }],
        "parameters": [{ "in": "path", "name": "sessionId", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Last assistant message interrupted and session counter updated" }, "404": { "description": "Session not found" } }
      }
    },
    "/api/goodspeech/v1/agents/sessions/{sessionId}/complete": {
      "post": {
        "summary": "Complete a persistent voice agent session",
        "description": "Closes the owned session, records its outcome, and queues an agent.session.completed webhook event.",
        "servers": [{ "url": "https://base.goodos.app" }],
        "security": [{ "GoodBaseSession": [] }, { "GoodOSBearer": [] }],
        "parameters": [{ "in": "path", "name": "sessionId", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "outcome": { "type": "string", "enum": ["resolved", "unresolved", "handed_off"] } } } } } },
        "responses": { "200": { "description": "Session completed" }, "404": { "description": "Session not found" } }
      }
    },
...
    "/api/goodspeech/v1/agents/{agentId}/tests/{testId}/run": {
      "post": {
        "summary": "Run a durable grounded agent regression test",
        "servers": [{ "url": "https://base.goodos.app" }],
        "security": [{ "GoodBaseSession": [] }, { "GoodOSBearer": [] }],
        "parameters": [{ "in": "path", "name": "agentId", "required": true, "schema": { "type": "string", "format": "uuid" } }, { "in": "path", "name": "testId", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Test result and captured response" }, "404": { "description": "Agent or test not found" } }
      }
    },
...
  }
}